I've created Callback DTOs.

+++ backend/MikartEnergy.Common/DTO/CallbackRequest/CallbackRequestDTO.cs
@@ -19,6 +19,6 @@ namespace MikartEnergy.Common.DTO.CallbackRequest
         public string AuthorPhone { get; set; }
         public string Message { get; set; }
         public string IntrerestedIn { get; set; }
-        public decimal Budget { get; set; }
+        public int Budget { get; set; }
     }
 }
