+++ backend/MikartEnergy.Common/DTO/CallbackRequest/NewCallbackRequestDTO.cs
@@ -13,6 +13,6 @@ namespace MikartEnergy.Common.DTO.CallbackRequest
         public string AuthorPhone { get; set; }
         public string Message { get; set; }
         public string IntrerestedIn { get; set; }
-        public decimal Budget { get; set; }
+        public int Budget { get; set; }
     }
 }
